% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_dataset.R
\name{compute_comparison}
\alias{compute_comparison}
\title{Compute comparison}
\usage{
compute_comparison(seuratObj, opt)
}
\arguments{
\item{seuratObj}{a Seurat object}

\item{opt}{a list of sub-lists (one for each individual comparison to compute),
with elements:
\itemize{
\item 'name', a character being the name that will appear in the app
\item 'subset', a list with \code{subset()} parameters
(except 'x', which is by default the Seurat object).
\item 'Ident', a character being the Idents of the Seurat object to use for
\code{Seurat::FindMarkers()}
\item 'FindMarkers', a list with \code{Seurat::FindMarkers()} parameters
(except 'object', which is by default the Seurat object).
The name of each element in the list is the name of parameters of the function,
and its value is the value to give to the parameters.
\item 'AggregateExpression', a list with \code{Seurat::AggregateExpression()}
parameters (except 'object').
The name of each element in the list is the name of parameters of the function,
and its value is the value to give to the parameters.
}}
}
\value{
a list of sub-lists (one for each individual marker table to compute).
Each sub-list is named with the 'name' given as input, and contains the elements:
\itemize{
\item 'table', the output of \code{Seurat::FindMarkers()}
\item 'aggrexpression', the output of \code{Seurat::AggregateExpression()}
}
}
\description{
Compute comparison tables.
}
\examples{
\dontrun{
data("small_seurat")
opt <- list(
list(
  name = "First comparison",
  subset = list(
    subset = substitute(CellType \%in\% c("DC"))
  ),
  Ident = "seurat_clusters",
  FindMarkers = list(ident.1 = 1),
  AggregateExpression = list(assays = "RNA",
                             slot = "scale.data",
                             group.by = c("seurat_clusters", "CellType")
  )
) #,
#list(
#   name = "Second comparison",
#   ...
#)
)
compute_comparison(small_seurat, opt)
}
}
