% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_dataset.R
\name{compute_markers}
\alias{compute_markers}
\title{Compute marker}
\usage{
compute_markers(seuratObj, opt)
}
\arguments{
\item{seuratObj}{a Seurat object}

\item{opt}{a list of sub-lists (one for each individual marker table to compute),
with elements:
\itemize{
\item 'name', a character being the name that will appear in the app
\item 'Ident', a character being the Idents of the Seurat object
to use for \code{Seurat::FindAllMarkers()}
\item 'FindAllMarkers', a list with \code{Seurat::FindAllMarkers()} parameters
(except 'object', which is by default the Seurat object).
The name of each element in the list is the name of parameters of the function,
and its value is the value to give to the parameters.
\item 'PseudobulkExpression', a list with \code{Seurat::PseudobulkExpression()}
parameters (except 'object', which is by default the Seurat object).
The name of each element in the list is the name of parameters of the function,
and its value is the value to give to the parameters.
}}
}
\value{
a list of sub-lists (one for each individual marker table to compute).
Each sub-list is named with the 'name' given as input, and contains the elements:
\itemize{
\item 'table', the output of \code{Seurat::FindAllMarkers()}
\item 'expression', the output of \code{Seurat::PseudobulkExpression()}
}
}
\description{
Compute several marker table and expression for
a dataset. The aim of this function is to facilitate creating a marker object
in the correct format, in order to add it to the database.
}
\examples{
\dontrun{
data("small_seurat")
opt <- list(
 list(
   name = "CellType markers",
   Ident = "CellType",
   FindAllMarkers = list(only.pos = TRUE,
                        min.pct = 0.25,
                        logfc.threshold = 0.5),
   PseudobulkExpression = list(assay = "RNA",
                             slot = "scale.data",
                             group.by = c("CellType", "orig.ident")
                         )
 ) #,
 # list(
 #  name = "seurat_clusters markers",
 #  #...
 # )
)
compute_markers(small_seurat, opt)
}
}
