% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_dataset.R
\name{populate_tiledb}
\alias{populate_tiledb}
\title{Populate TileDB-SOMA with a new dataset.}
\usage{
populate_tiledb(dir, force = F)
}
\arguments{
\item{dir}{a character, corresponding to the path folder of the dataset.
It is where a object.rds file is expected, and where the tiledb folder will be created.}

\item{force}{a boolean, if a tiledb folder already exists, should it be removed?}
}
\value{
None. It creates a TileDB-SOMA in a specified folder.
}
\description{
Populate TileDB-SOMA with a new dataset.
}
