% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = "url",
  uiPattern = "/",
  studies = NULL,
  cache_path = NULL,
  log_path = "log",
  authr_file = NULL,
  tabs = list(),
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{studies}{a data.frame summarizing of all available studies.
The data.frame have as colnames:
\code{c("title", "description", "doi", "date", "nsamples", "nfeatures", "ncells", "output", "rds")}.
By default the app will be run on a very small subset
of pbmc3k present in the package. Check out
\code{vignette("database", package = "singlecellviz")} for more information.}

\item{cache_path}{a path to the cache directory to use.
If the folder already exists, make sure that the use has read and write access.}

\item{log_path}{a path to the log directory, where a \code{telemetry.txt} file will
be created. If the folder or file already exists, make sure that the use has read and write access.}

\item{authr_file}{a path to the authr .txt file, where the columns
\code{user}, \code{password_hash}, and \code{permissions} are available.
If the folder or file already exists, make sure that the use has read and write access.}

\item{tabs}{a list with names being tabs among
\code{['homepage', 'information', 'explore', 'markers', 'differential', 'download']}
and values being a boolean (TRUE to show the tab, FALSE otherwise). By default all tabs are shown}

\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
Run the Shiny Application
}
