% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_seurat}
\alias{small_seurat}
\title{A subset of data from pbmc3k}
\format{
## `small_seurat`
A Seurat object
\describe{
A very small subset of the pbmc3k for testing purpose.
It is the element 'seuratObj'
in the object under inst/extdata/small_example/small_example.rds
}
}
\source{
\url{https://github.com/satijalab/seurat-data}
}
\usage{
small_seurat
}
\description{
A subset of data from pbmc3k
}
\keyword{datasets}
