% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_dataset.R
\name{populate_tiledb}
\alias{populate_tiledb}
\title{Populate TileDB-SOMA with a new dataset.}
\usage{
populate_tiledb(yaml)
}
\arguments{
\item{yaml}{a character, corresponding to the path file of a
config.yaml file for a dataset. Check out the section 'Adding a new dataset'
in \code{vignette("database", package = "singlecellviz")}
for a description of the config.yaml file.}
}
\value{
None. It creates a TileDB-SOMA in a specified folder.
}
\description{
Populate TileDB-SOMA with a new dataset.
}
