% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_dataset.R
\name{update_summary}
\alias{update_summary}
\title{Update dataset summary}
\usage{
update_summary(db_dir, output_file, config_filename = "config.yaml")
}
\arguments{
\item{db_dir}{a character, corresponding to the path of the database
(containing subfolders)}

\item{output_file}{a character, corresponding to the path and filename
to give to the txt file created}

\item{config_filename}{a character, corresponding to the name of the
yaml configuration files to look for (by default  \code{"config.yaml"})}
}
\value{
None. It creates a txt file.
}
\description{
Creates a txt file summarizing information about
all available datasets by automatically
searching for \code{config.yaml} files in each subfolder of specified
\code{db_dir}.
}
\examples{
\dontrun{
update_summary("db/", "dataset_summary.txt")
}
}
